Implementando a funcionalidade de atualizar Estabelecimento

+++ desafio/src/main/java/br/com/estacionamento/model/Estabelecimento.java
@@ -1,6 +1,10 @@
 package br.com.estacionamento.model;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -17,7 +21,7 @@ public class Estabelecimento {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     @NotNull(message = "Este campo é de preenchimento obrigatório")
     private String nome;
     @NotNull(message = "Este campo é de preenchimento obrigatório")
